Factory 에서 TType을 사용하여 분류하게 처리함

+++ gugu/Assets/02.Scripts/Actor/Actor/Actor.cs
@@ -54,7 +54,7 @@ public class Actor : PoolingObject
     }
     protected override void ReturnToPool()
     {
-        ActorManager.Instance.RegisterActorPool(worldID,(uint)type,pathHashCode);
+        ActorManager.Instance.RegisterActorPool(worldID,type,pathHashCode);
     }
     #endregion
